+++ examples/001-flutter_default_counter_app/lib/pages/async_counter_app.dart
@@ -31,7 +31,7 @@ class MyHomePage extends StatelessWidget {
               onError: (error) => Text(counterRM.error.message),
               onData: (data) => Text(
                 '$data',
-                style: Theme.of(context).textTheme.headline4,
+                style: Theme.of(context).textTheme.headline,
               ),
             ),
           ],