+++ examples/001-flutter_default_counter_app/lib/pages/simple_counter_with_error.dart
@@ -30,7 +30,7 @@ class MyHomePage extends StatelessWidget {
                   return Text(
                     //get the current value of the counter
                     '${counterRM.value}',
-                    style: Theme.of(context).textTheme.headline4,
+                    style: Theme.of(context).textTheme.headline,
                   );
                 }),
           ],