+++ examples/001-flutter_default_counter_app/test/async_counter_app_test.dart
@@ -1,0 +1,17 @@
+//Can not be tested because the business logic is mixed with the UI logic
+
+// The Random().nextBool() can not be neither expected nor mocked.
+
+// onPressed: () {
+//   counterRM.setValue(
+//     () async {
+//       await Future.delayed(Duration(seconds: 1));
+//       if (Random().nextBool()) {
+//         throw Exception('A Counter Error');
+//       }
+//       return counterRM.value + 1;
+//     },
+//     catchError: true,
+//   );
+// },
+void main() {}