+++ examples/002-form_validation_with_reactive_model/test/widget_test.dart
@@ -1,30 +1,96 @@
-// This is a basic Flutter widget test.
-//
-// To perform an interaction with a widget in your test, use the WidgetTester
-// utility that Flutter provides. For example, you can send tap and scroll
-// gestures. You can also use WidgetTester to find child widgets in the widget
-// tree, read text, and verify that the values of widget properties are correct.
-
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'package:form_validation_with_reactive_model/main.dart';
 
 void main() {
-  testWidgets('Counter increments smoke test', (WidgetTester tester) async {
+  Finder emailTextField;
+  Finder passwordTextField;
+  Finder activeLoginButton;
+  setUp(
+    () {
+      emailTextField = find.byWidgetPredicate(
+        (widget) =>
+            widget is TextField &&
+            widget.decoration.labelText == "Email Address",
+      );
+
+      passwordTextField = find.byWidgetPredicate(
+        (widget) =>
+            widget is TextField && widget.decoration.labelText == "Password",
+      );
+
+      //active login button is that with a non null onPressed parameter
+      activeLoginButton = find.byWidgetPredicate(
+        (widget) => widget is RaisedButton && widget.onPressed != null,
+      );
+    },
+  );
+
+  testWidgets('email validation', (WidgetTester tester) async {
     // Build our app and trigger a frame.
     await tester.pumpWidget(MyApp());
 
-    // Verify that our counter starts at 0.
-    expect(find.text('0'), findsOneWidget);
-    expect(find.text('1'), findsNothing);
+    expect(find.text('Enter a valid Email'), findsNothing);
+    expect(find.text('Email : Enter a valid Email'), findsNothing);
 
-    // Tap the '+' icon and trigger a frame.
-    await tester.tap(find.byIcon(Icons.add));
+    //Non valid Email
+    await tester.enterText(emailTextField, 'mail');
     await tester.pump();
+    expect(find.text('Enter a valid Email'), findsOneWidget);
+    expect(find.text('Email : Enter a valid Email'), findsOneWidget);
 
-    // Verify that our counter has incremented.
-    expect(find.text('0'), findsNothing);
-    expect(find.text('1'), findsOneWidget);
+    //valid Email
+    await tester.enterText(emailTextField, 'mail@');
+    await tester.pump();
+    expect(find.text('Enter a valid Email'), findsNothing);
+    expect(find.text('Email : mail@'), findsOneWidget);
+  });
+
+  testWidgets('password validation', (WidgetTester tester) async {
+    await tester.pumpWidget(MyApp());
+
+    expect(find.text('Enter a valid password'), findsNothing);
+    expect(find.text('password : Enter a valid password'), findsNothing);
+
+    //Non valid password
+    await tester.enterText(passwordTextField, 'pas');
+    await tester.pump();
+    expect(find.text('Enter a valid password'), findsOneWidget);
+    expect(find.text('password : Enter a valid password'), findsOneWidget);
+
+    //valid password
+    await tester.enterText(passwordTextField, 'password');
+    await tester.pump();
+    expect(find.text('Enter a valid password'), findsNothing);
+    expect(find.text('password : password'), findsOneWidget);
+  });
+
+  testWidgets('active login button', (WidgetTester tester) async {
+    await tester.pumpWidget(MyApp());
+
+    //before tapping login button is inactive
+    expect(activeLoginButton, findsNothing);
+
+    //Non valid email and valid password
+    await tester.enterText(emailTextField, 'mail');
+    await tester.enterText(passwordTextField, 'password');
+    await tester.pump();
+    //login button is inactive
+    expect(activeLoginButton, findsNothing);
+
+    //valid email and non valid password
+    await tester.enterText(emailTextField, 'mail@');
+    await tester.enterText(passwordTextField, 'pa');
+    await tester.pump();
+    //login button is inactive
+    expect(activeLoginButton, findsNothing);
+
+    //valid email and password
+    await tester.enterText(emailTextField, 'mail@');
+    await tester.enterText(passwordTextField, 'password');
+    await tester.pump();
+    //login button is active
+    expect(activeLoginButton, findsOneWidget);
   });
 }