+++ examples/003-async_counter_app_with_injector/lib/counter_with_injector.dart
@@ -19,39 +19,44 @@ class CounterStore {
 }
 
 class MyApp extends StatelessWidget {
-  final counterRM = ReactiveModel.create(CounterStore(0));
+  //Remove this line
+  //final counterRM = ReactiveModel.create(CounterStore(0));
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Flutter Demo',
-      theme: ThemeData(
-        primarySwatch: Colors.blue,
-      ),
-      home: _MyScaffold(
-        counterRM: counterRM,
-      ),
-    );
+    return Injector(
+        inject: [Inject(() => CounterStore(0))],
+        builder: (context) {
+          return MaterialApp(
+            title: 'Flutter Demo',
+            theme: ThemeData(
+              primarySwatch: Colors.blue,
+            ),
+            home: _MyScaffold(),
+          );
+        });
   }
 }
 
 class _MyScaffold extends StatelessWidget {
-  final ReactiveModel<CounterStore> counterRM;
-
-  const _MyScaffold({Key key, this.counterRM}) : super(key: key);
+  //Remove this line
+  // final ReactiveModel<CounterStore> counterRM;
+  // const _MyScaffold({Key key, this.counterRM}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    return MyHomePage(
-      title: 'Flutter Demo Home Page',
-      counterRM: counterRM,
-    );
+    return MyHomePage(title: 'Flutter Demo Home Page');
   }
 }
 
 class MyHomePage extends StatelessWidget {
-  MyHomePage({Key key, this.title, this.counterRM}) : super(key: key);
+  MyHomePage({Key key, this.title}) : super(key: key);
   final String title;
-  final ReactiveModel<CounterStore> counterRM;
+
+  final ReactiveModel<CounterStore> counterRM = ReactiveModel<CounterStore>();
+
+  //You can use Injector.getAsReactive
+  // final ReactiveModel<CounterStore> counterRM =
+  //     Injector.getAsReactive<CounterStore>();
 
   @override
   Widget build(BuildContext context) {
@@ -74,7 +79,7 @@ class MyHomePage extends StatelessWidget {
               onError: (error) => Text(counterRM.error.message),
               onData: (data) => Text(
                 '${data.count}',
-                style: Theme.of(context).textTheme.headline4,
+                style: Theme.of(context).textTheme.headline,
               ),
             ),
           ],