+++ examples/005-double_async_counter_with_error/lib/counter_page_3.dart
@@ -84,7 +84,7 @@ class CounterPage extends StatelessWidget {
               style: TextStyle(color: Colors.red),
             ),
             onData: (data) => Text(
-              ' ${counterService.state.counter.count}',
+              '${counterService.state.counter.count}',
               style: TextStyle(fontSize: 30),
             ),
           ),