+++ examples/008-clean_architecture_firebase_login/lib/main.dart
@@ -35,6 +35,7 @@ class MyApp extends StatelessWidget {
           appleCheckerRM.setState((s) => s.check());
           userServiceRM.setState((s) => s.currentUser());
         },
+        onIdle: () => Container(),
         onWaiting: () => SplashScreen(),
         onError: (error) => Text(error.toString()),
         onData: (_) {